requirements.txt
flask==2.3.3
Flask-Markdown
Flask-SocketIO
pyyaml